% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_a_method_indel_information_to_master_table}
\alias{add_a_method_indel_information_to_master_table}
\title{Add column that states whether the variant called by a method is an indel or not}
\usage{
add_a_method_indel_information_to_master_table(
  input_table,
  vcf_file,
  method_name,
  truth_name
)
}
\arguments{
\item{input_table}{A data.frame. The master table to add the new column.}

\item{vcf_file}{1-length string. The address of the method VCF file.}

\item{method_name}{1-length string. The name of the method.}

\item{truth_name}{1-length string. The name of the ground-truth.}
}
\value{
A data.frame
}
\description{
Add column that states whether the variant called by a method is an indel or not
}
