% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_format_tag_from_vcf}
\alias{add_format_tag_from_vcf}
\title{Add a FORMAT tag from a VCF file to a master table}
\usage{
add_format_tag_from_vcf(input_table, method_name, tag, vcf_file)
}
\arguments{
\item{input_table}{A data.frame. The input master table.}

\item{method_name}{A 1-length string. The name of the method.}

\item{tag}{A 1-length string. The name of the tag to add. It must be exactly
how the tag is written in the VCF file (case sensitive).}

\item{vcf_file}{A 1-length string. The path to the VCF file from which the tag
is extracted.}
}
\value{
A data.frame.
}
\description{
The function adds to `input_table` the new column <tagName>_<methodName>,
  where <tagName> is `tag` in lower case and <methodName> is `method_name`.
}
