% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_method_vs_truth_comparison_to_master_table}
\alias{add_method_vs_truth_comparison_to_master_table}
\title{Add classiffication of variants calls by comparing to the ground-truth to a
  master table}
\usage{
add_method_vs_truth_comparison_to_master_table(
  input_table,
  method_name,
  truth_name,
  replace_column = FALSE
)
}
\arguments{
\item{input_table}{A data.frame. The master table to add the new column.}

\item{method_name}{A 1-length string. The name of the method to compare to the
ground-truth.}

\item{truth_name}{A 1-length string. The name of the ground-truth.}

\item{replace_column}{1-length boolean (default is FALSE). If the classification
column already exists and it is desired to replace it.}
}
\value{
A data.frame.
}
\description{
Compare variants calls of a method to the ground-truth and classify each one
  as true-positiove (TP), false-negative (FN), false-positive (FP), or
  true-negative (TN).
}
