% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_number_of_n_cigar_reads}
\alias{add_number_of_n_cigar_reads}
\title{Add the count of N-cigar reads to the master table}
\usage{
add_number_of_n_cigar_reads(input_table, input_bam, dataset_name = NULL)
}
\arguments{
\item{input_table}{A data.frame. The master table to add the new column.}

\item{input_bam}{A `GenomicAlignments` object from which N-cigar-read counts
are got.}

\item{dataset_name}{A 1-length string used to set the name of the new column.}
}
\value{
A data.frame.
}
\description{
This function adds a new column to a master table with the counts of
N-cigar reads that overlap each site.
}
