% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_read_coverage_from_bam_to_master_table}
\alias{add_read_coverage_from_bam_to_master_table}
\title{Add read coverage (taken from BAM file) of each dataset to a master table}
\usage{
add_read_coverage_from_bam_to_master_table(
  input_table,
  dataset_coverage,
  dataset_name
)
}
\arguments{
\item{input_table}{A master table to add read-coverage columns.}

\item{dataset_coverage}{Object generated by \code{IRanges::coverage} function.}

\item{dataset_name}{A 1-length string. Name of the dataset used to genere 
`dataset_coverage`.}
}
\value{
A data frame.
}
\description{
Add read coverage (taken from BAM file) of each dataset to a master table
}
