% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_some_read_statistic_from_bam_metacolumn}
\alias{add_some_read_statistic_from_bam_metacolumn}
\title{Add some statistic of a tag about reads of a BAM file}
\usage{
add_some_read_statistic_from_bam_metacolumn(
  input_table,
  col_name,
  galn,
  tag,
  stat
)
}
\arguments{
\item{input_table}{A data.frame. The input master table.}

\item{col_name}{A 1-length string. The name of the new column to add.}

\item{galn}{A GAlignment object. The alingnment from where the tag is 
extracted.}

\item{tag}{A 1-length string. The name of tag that is wanted to calculate
the statistic `stat`. The name must be exactly the same of one column
name from `mcols(galn)`.}

\item{stat}{A function used to calculate the desired statistic.}
}
\value{
A data.frame.
}
\description{
Add some statistic of a tag about reads of a BAM file
}
