% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_splice_site_info_to_master_table}
\alias{add_splice_site_info_to_master_table}
\title{Add columns about spice sites to a master table}
\usage{
add_splice_site_info_to_master_table(
  input_table,
  splice_sites,
  max_dist_from_splice_site = 20,
  multiply_max_dist = 100
)
}
\arguments{
\item{input_table}{A data.frame generated by \code{initiate_master_table}.}

\item{splice_sites}{A data.frame generated by \code{get_splice_sites_info}.}

\item{max_dist_from_splice_site}{A 1-length integer (default is 20) that
indicates the the maximum distance of a varaint from any splice site to
consider it a variant near a splice site.}

\item{multiply_max_dist}{A 1-length integer (default is 100). To create the
column `ss_shortest_dist` from column `ss_dist`, NA values (which means
variables far from any splice site) are converted to integers. To make
these integers be a high value, they are set to
\code{max_dist_from_splice_site * multiply_max_dist}.}
}
\value{
A data.frame.
}
\description{
Add columns about spice sites to a master table
}
