% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_the_ground_truth_indel_information_to_master_table}
\alias{add_the_ground_truth_indel_information_to_master_table}
\title{Add column that states whether the variant in the ground-truth is an indel or not}
\usage{
add_the_ground_truth_indel_information_to_master_table(
  input_table,
  vcf_file,
  truth_name
)
}
\arguments{
\item{input_table}{A data.frame. The master table to add the new column.}

\item{vcf_file}{1-length string. The address of the ground-truth VCF file.}

\item{truth_name}{1-length string. The name of the ground-truth.}
}
\value{
A data.frame
}
\description{
Add column that states whether the variant in the ground-truth is an indel or not
}
