% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_two_method_comparison_to_master_table}
\alias{add_two_method_comparison_to_master_table}
\title{Add column to compare two methods in a master table}
\usage{
add_two_method_comparison_to_master_table(
  input_table,
  method1_name,
  method2_name
)
}
\arguments{
\item{input_table}{A data.frame. The master table to add the new column.}

\item{method1_name}{A 1-length string. The name of the first method.}

\item{method2_name}{A 1-length string. The name of the second method.}
}
\value{
A data.frame
}
\description{
This function adds a column in a master table that compares two given methods.
The new column informs whether the variant was called only by the first method
  (`method1_name`), only by the second method (`method2_name`), by both
  methods (`"both"`), or neither of them (`"neither"`).
}
