% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_variant_type_to_master_table}
\alias{add_variant_type_to_master_table}
\title{Add a column that states the variant type according a method or the ground truth}
\usage{
add_variant_type_to_master_table(input_table, vcf_file, method_name)
}
\arguments{
\item{input_table}{A data.frame. The master table to add the new column.}

\item{vcf_file}{1-length string. The address of the ground-truth VCF file.}

\item{method_name}{1-length string. The name of the ground-truth.}
}
\value{
A data.frame
}
\description{
The new column stores the values "snp", "insertion", or "deletion",
  respectively to the variant type. If the variant type can not be defined --
  because the variant is heterozygous alternative and the alleles are a mix
  between SNP, insertion, or deletion -- the value returned is "mix".
}
