% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_accuracy_measures}
\alias{calc_accuracy_measures}
\title{Calculate precision, sensitivity and F1-score of a method in a master table}
\usage{
calc_accuracy_measures(input_table, method_name, truth_name)
}
\arguments{
\item{input_table}{A data.frame. The master table.}

\item{method_name}{A 1-length string. The name of the method to calculate the
accuracy measures.}

\item{truth_name}{A 1-length string. The name of the ground-truth to validate
the method.}
}
\value{
A named vector with the accuracy measures.
}
\description{
Calculate precision, sensitivity and F1-score of a method in a master table
}
