% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_precision_recall_for_coverage_ranges}
\alias{calculate_precision_recall_for_coverage_ranges}
\title{Calculate accuracy measures, per read coverage ranges, for multiple master
  tables, and output number of true variant per range}
\usage{
calculate_precision_recall_for_coverage_ranges(
  ...,
  experiment_names,
  method_names,
  output_method_names = NULL,
  data_names,
  truth_names,
  coverage_ranges,
  what
)
}
\arguments{
\item{...}{Master tables.}

\item{experiment_names}{Vector of strings. Name of the experiments in the same 
order of the master tables.}

\item{method_names}{Vector or list of strings. Names of the methods to compare.
If all experiments compare the same methods, `method_names` can be a single
vector. Otherwise, `method_names` must be a list, in which each element 
specifies the names of the methods to compare for each experiment.}

\item{output_method_names}{NULL (default), or a vector or a list of strings.
Names of the methods to output.}

\item{data_names}{Vector or list of strings.}

\item{truth_names}{Vector or list of strings.}

\item{coverage_ranges}{Data.frame of two columns. The first column stores the
begining of the ranges and the second the end.}

\item{what}{Vector or list of strings.}
}
\value{
A data.frame.
}
\description{
This is an alternative function for `calculate_precision_recall_for_multi_master_tables2`.
  It differs from the latter by using ranges of read coverage, i.e., there are
  maximums.
}
