% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_precision_recall_for_multi_master_tables}
\alias{calculate_precision_recall_for_multi_master_tables}
\title{Calculate accuracy measures for multiple master tables}
\usage{
calculate_precision_recall_for_multi_master_tables(
  ...,
  experiment_names,
  method_names,
  output_method_names = NULL,
  data_names,
  truth_names,
  coverage_thresholds = c(3, 15, 40, 100),
  what
)
}
\arguments{
\item{...}{Master tables.}

\item{experiment_names}{Vector of strings. Name of the experiments in the same 
order of the master tables.}

\item{method_names}{Vector or list of strings. Names of the methods to compare.
If all experiments compare the same methods, `method_names` can be a single
vector. Otherwise, `method_names` must be a list, in which each element 
specifies the names of the methods to compare for each experiment.}

\item{output_method_names}{NULL (default), or a vector or a list of strings.
Names of the methods to output.}

\item{data_names}{Vector or list of strings.}

\item{truth_names}{Vector or list of strings.}

\item{coverage_thresholds}{Vector or list of integers.}

\item{what}{Vector or list of strings.}
}
\value{
A data.frame.
}
\description{
Create a data.frame with accuracy measures for multiple master tables.
}
