% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_accuracy_per_coverage}
\alias{check_accuracy_per_coverage}
\title{Barplot of accuracy measures for different methods and coverage}
\usage{
check_accuracy_per_coverage(
  master_table,
  method_names,
  data_name,
  truth_name,
  interval_start,
  interval_end
)
}
\arguments{
\item{master_table}{A data.frame. The input master table.}

\item{method_names}{A vector of strings. The name of the methods to be compared.}

\item{data_name}{A 1-length string. The name of the dataset used with the methods 
to be compared.}

\item{truth_name}{A 1-length string. The name of the ground-truth.}

\item{interval_start}{A vector of integers. The start position of all intervals.}

\item{interval_end}{A vector of integers. The end position of all intervals.}
}
\value{
A ggplot object.
}
\description{
Return a ggplot object for visualization.
}
