% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagCorrection.R
\name{flagCorrection}
\alias{flagCorrection}
\title{Make the flags of a BAM be the same as they were before SplitNCigarReads}
\usage{
flagCorrection(input_bam, input_sncr_bam, output_bam, threads)
}
\arguments{
\item{input_bam}{A 1-length string. Path of the BAM file before
SplitNCigarReads.}

\item{input_sncr_bam}{A 1-length string. Path of the BAM file after
SplitNCigarReads.}

\item{output_bam}{A 1-length string. File path to save the 
output BAM file with corrected flags.}

\item{threads}{A 1-length integer. Number of threads.}
}
\value{
NULL.
}
\description{
This function is specially important to be used combined with DeepVariant
  to call variants from Iso-Seq data.
}
