% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{get_splice_sites_info2}
\alias{get_splice_sites_info2}
\title{Get all splice-sites and read-end positions of a BAM files}
\usage{
get_splice_sites_info2(input_bam, threads)
}
\arguments{
\item{input_bam}{The input BAM file to extract splice site positions from.}

\item{threads}{Number of threads.}
}
\value{
A 2-element list (`splice_site`, and `transcript_end`). Element
  `splice_site` should contain the same returned by \code{get_splice_sites_info}.
  Element `transcript_end` counts of read ends (start or final/end site of the read).
}
\description{
Similar to \code{get_splice_sites_info}, but also count read ends.
}
\details{
This function may take several hours to run.
}
