% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{get_splice_sites_information}
\alias{get_splice_sites_information}
\title{Get all splice sites positions of a BAM files}
\usage{
get_splice_sites_information(input_bam, threads)
}
\arguments{
\item{input_bam}{The input BAM file to extract splice site positions from.}

\item{threads}{Number of threads.}
}
\value{
A data.frame in which each row is a splice-site position.
}
\description{
Get the position of all splice sites in a BAM file and the number of reads that
support each one of them. Besides that, indicate whether the splice site is
acceptor or donor site.
}
\details{
This function may take several hours to run.
}
