% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gt_vt_method}
\alias{gt_vt_method}
\title{get the genotype and the variant type of the calls of a primary and 
secundary method}
\usage{
gt_vt_method(input_table, gt_first, gt_second, vt_first, vt_second)
}
\arguments{
\item{input_table}{A data.frame. The input master table.}

\item{gt_first}{A 1-length string. The name of the column that stores the
genotypes called by a method. This method is called the first method.}

\item{gt_second}{A 1-length string. If the first method does not call some 
variants, extract their genotype from column `gt_second`.}

\item{vt_first}{A 1-length string. The name of the column that stores the
variant type called by a method. This method is called the first method.}

\item{vt_second}{A 1-length string. If the first method does not call some 
variants, extract their variant type from column `vt_second`.}
}
\value{
A data.frame.
}
\description{
Create a new data.frame from a input master table. The rows correspond to
  the same variants in `input_table` and in the same order.
}
