% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{is_indel_method}
\alias{is_indel_method}
\title{Add a column to state whether the variant is an indel or not}
\usage{
is_indel_method(input_table, first_method, second_method = NULL)
}
\arguments{
\item{input_table}{A data.frame. The master table to add the new column.}

\item{first_method}{A 1-length string. The name of the first method.}

\item{second_method}{A 1-length string. The name of the second method}
}
\value{
A data.frame.
}
\description{
To classify the variant as an indel, first the function looks at the VCF file
of the `first_method`. If the variant is not there, if looks at the VCF file
of the `second_method`.
}
\details{
Output meaning:
  * `-1` means that the variant type couldn't be defined, because it is 
    heterozygous alternative;
  * `0` means it is not an indel;
  * `1` means it is an indles;
  * `NA` means the `first_method` (and the `second_method`) didn't call the
    variant.
}
