% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jurkat_master_table_v6_notFiltered}
\alias{jurkat_master_table_v6_notFiltered}
\title{The jurkat_master_table_v6_notFiltered master table}
\format{
A data.frame with dimensions equal to (1003676, 26).

Each row is a variant called by at least one of the methods. For column
descriptions, see \url{https://docs.google.com/document/d/1REIXRS6_oGEjTtlokwZ0pBQmoZcEE7L6Nkq-nCMoI5c/edit}
}
\usage{
data(jurkat_master_table_v6_notFiltered)
}
\description{
A master table generated from Jurkat human cells. Variant callers to compare
are DeepVariant and GATK on Iso-Seq data. The ground-truth was generated with
GATK from short-read data.
}
\details{
Updates from `jurkat_master_table_v6_notFiltered`:
  * DeepVariant calls filtered by QUAL>=20;
  * VCF tag QD of the gound-truth was added as column `qd_trDnaMerged`.

Methods to avaluate included: dv_s_fc_q20, gatk_s.

Ground-truth: tr_dna_merged.
}
\keyword{datasets}
