% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jurkat_master_table_v7_notFiltered}
\alias{jurkat_master_table_v7_notFiltered}
\title{The jurkat_master_table_v7_notFiltered master table}
\format{
A data.frame with dimensions equal to (1018480, 32).
}
\usage{
data(jurkat_master_table_v7_notFiltered)
}
\description{
A master table generated from Jurkat human cells. Variant callers to compare
are DeepVariant and GATK on Iso-Seq data. The ground-truth was generated with
GATK from short-read data.
}
\details{
Each row is a variant called by at least one of the methods. For column
descriptions, see \url{https://docs.google.com/document/d/1REIXRS6_oGEjTtlokwZ0pBQmoZcEE7L6Nkq-nCMoI5c/edit}

Methods to evaluate: dv_s_fc, gatk_s.

Ground-truth: tr_dna_merged.

No filtering applied.
}
\keyword{datasets}
