% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{method_homopolymer_indels}
\alias{method_homopolymer_indels}
\title{Get indels in homopolymers}
\usage{
method_homopolymer_indels(
  input_table,
  first_method_name,
  second_method_name,
  vcf_first,
  vcf_second,
  method_dataset_name,
  homopolymers,
  ref_fasta_seqs,
  min_isoseq_coverage,
  genotyped_alt
)
}
\arguments{
\item{input_table}{A data.frame. The input master table.}

\item{first_method_name}{A 1-length string of the name of a method. First,
the variant information that is taken is related to this method. If a
variant is not called by the method, its information is taken relating
to the method specifiec in `second_method_name`.}

\item{second_method_name}{A 1-length string. The name of a method.}

\item{vcf_first}{A 1-length string. The path of the VCF file of the first
method.}

\item{vcf_second}{A 1-length string. The path of the VCF file of the
second method.}

\item{method_dataset_name}{A 1-length string. Name of the dataset used to
call variants by the methods to be compared.}

\item{homopolymers}{A CompressedIRangesList object. It should store all 
homopolymers, it's nucleotive types and lengths, of the genome used as
the reference to call the variants. It is gerated by the function
`sarlacc::homopolymerFinder`.}

\item{ref_fasta_seqs}{A DNAStringSet object. The sequences of the genome
used as the reference to call the variants. It's names must be in the
form like "chr1", "chr2", ..., "chrX", "chrY".}

\item{min_isoseq_coverage}{min iso-seq read coverage to filter.}

\item{genotyped_alt}{One of the strings "find" or "same". If "find", the
function finds the correct alternative allele based on the genotype. This
option must be chosen if there are multiple values for column ALT, which
is the case of VCF files output by DeepVariant and Clair3. Since GATK's
VCF files keep only the genotyped alternative allele in column ALT, this
argument should be "same".}
}
\value{
A data.frame.
}
\description{
Create a new data.frame that stores deletions of inside homopolymers, and
insertions of a same nucleotide type that happen inside homopolymers of
that same nucleotide type. Homopolymer length equal to 1 means
non-homopolymers.
}
