% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{precision_recall_curve_per_coverage}
\alias{precision_recall_curve_per_coverage}
\title{Precision-recall curves to compare methods, using different filtering on read
  coverage}
\usage{
precision_recall_curve_per_coverage(
  master_table,
  method_names,
  output_method_names = NULL,
  data_name,
  truth_name,
  coverage_threshold,
  indels_snps_separated
)
}
\arguments{
\item{master_table}{A data.frame. The input master table.}

\item{method_names}{A vector of strings. The names of the methods to be compared.}

\item{output_method_names}{A vector of strings. How to output the names of the
methods to be compared. The default is NULL.}

\item{data_name}{A 1-length string. The name of the dataset used with the methods 
to be compared.}

\item{truth_name}{A 1-length string. The name of the ground-truth.}

\item{coverage_threshold}{A vector of integers. The minimum thresholds to filer 
by read coverage. Each element defines a point in the precision-recall curves.}

\item{indels_snps_separated}{A 1-length logical vector. Default is equal to FALSE.
If TRUE, each precision-recall curves is separated into two: one for indels
and other for SNPs.}
}
\value{
A ggplot object.
}
\description{
Precision-recall curves to compare methods, using different filtering on read
  coverage
}
