% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splice_junction_analysis_table}
\alias{splice_junction_analysis_table}
\title{Generate data for splice-junction analysis using multiple master tables}
\usage{
splice_junction_analysis_table(
  ...,
  experiment_names,
  truth_names,
  method_dataset_name,
  method_names,
  output_method_names,
  variant_type,
  min_isoseq_coverage
)
}
\arguments{
\item{...}{Data.frames. Each data.frame is a master table.}

\item{experiment_names}{A vector of strings with length equal to the number of master
tables input in `...`. Names of the datasets for each data table.}

\item{truth_names}{A vector of strings. Names of the ground truth in each data table
especified in `...`.}

\item{method_dataset_name}{A vector of strings. Name of the datasets used to call
variants, by the methods to be compared, in each input master table.}

\item{method_names}{A vector of strings. The name of the methods to be compared defined in
the input master tables. All master table must have the same `method_names`.}

\item{output_method_names}{A vector of strings. The output name of the methods to be
compared defined in the input master tables. The output name of the methods will be
the same for all master table.}

\item{variant_type}{A 1-length string. Possible values are "snp" or "indel".}

\item{min_isoseq_coverage}{A 1-length integer. The threshod value for the minimum
Iso-Seq read coverage.}
}
\value{
A list of two elements (`acc_sj` and `n_test`) to be used to draw a chart to
 analyse the variant-calling performance of sites near to and far from splice junctions.
 `acc_sj` stores the calculated performance measures, and `n_test` stores the number of
 observed variants in each case.
}
\description{
Generate data to plot variant performance of sites near to and far from splice junctions
  comparisons using multiple master table as facets.
}
