% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_genotype}
\alias{standardize_genotype}
\title{Standardize genotypes}
\usage{
standardize_genotype(gt)
}
\arguments{
\item{gt}{A vector of strings. The input genotypes.}
}
\value{
A vector of strings.
}
\description{
This function rewrite genotype like:
* 0|1 to 0/1;
* 0/2, 0/3, ..., to 0/1;
* 2/2, 3/3, ..., to 1/1;
* 1/3, 2/3, ..., to 1/2
}
