% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_format_tag_from_vcf}
\alias{add_format_tag_from_vcf}
\title{Add a FORMAT tag from a VCF file to a master table}
\usage{
add_format_tag_from_vcf(input_table, col_name, tag, vcf_file)
}
\arguments{
\item{input_table}{A data.frame. The input master table.}

\item{col_name}{A 1-length string. the name of the column to add.}

\item{tag}{A 1-length string. The name of the tag to add. It must be exactly
how the tag is written in the VCF file (case sensitive).}

\item{vcf_file}{A 1-length string. The path to the VCF file from which the tag
is extracted.}
}
\value{
A data.frame.
}
\description{
If a variant in the VCF file doesn't contain the specified tag, the function
return -1 for that variant.
}
