% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{add_splice_site_info_to_master_table}
\alias{add_splice_site_info_to_master_table}
\title{Add columns about spice sites the a master table}
\usage{
add_splice_site_info_to_master_table(
  input_table,
  splice_sites,
  max_dist_from_splice_site = 20
)
}
\arguments{
\item{input_table}{A data.frame generated by \code{initiate_master_table}.}

\item{splice_sites}{A data.frame generated by \code{get_splice_sites_info}.}

\item{max_dist_from_splice_site}{A 1-length integer that indicates the
the maximum distance of a varaint from any splice site to consider it
a variant near a splice site.}
}
\value{
A data.frame.
}
\description{
Add columns about spice sites the a master table
}
