% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_accuracy_per_coverage}
\alias{check_accuracy_per_coverage}
\title{Barplot of accuracy measures for different methods and coverage}
\usage{
check_accuracy_per_coverage(
  master_table,
  method_names,
  output_method_names = NULL,
  data_name,
  truth_name,
  coverage_threshold
)
}
\arguments{
\item{master_table}{A data.frame. The input master table.}

\item{method_names}{A vector of strings. The names of the methods to be compared
in the master table.}

\item{output_method_names}{A vector of strings. How to output the names of the
methods to be compared. The defaut is NULL.}

\item{data_name}{A 1-length string. The name of the dataset used with the methods 
to be compared.}

\item{truth_name}{A 1-length string. The name of the ground-truth.}

\item{coverage_threshold}{A vector of integers. The minimum thresholds to filer 
by read coverage.}
}
\value{
A ggplot object.
}
\description{
Return a ggplot object for visualization.
}
