% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{igv_batch_screenshots}
\alias{igv_batch_screenshots}
\title{Generate igv batch screenshots script}
\usage{
igv_batch_screenshots(
  chrm,
  pos,
  output_dir,
  prefix,
  snapshot_path,
  windows_size = 1501,
  screenshot_number = 100,
  output_positions = FALSE
)
}
\arguments{
\item{chrm}{A vector of string. Chromossome name, e.g. "chrm1".}

\item{pos}{A vector of integers. Positions to center the screenshot.}

\item{output_dir}{A 1-length string. Directory where the IGV-batch-screenshot 
script is saved.}

\item{prefix}{A 1-length string.}

\item{snapshot_path}{A 1-length string. Directory where the IGV-batch-screenshot 
script will save the screenshots.}

\item{windows_size}{1-length integer. Window size of the screenshot.}

\item{screenshot_number}{1-length integer. Number of screenshot to generate.}

\item{output_positions}{If TRUE, output a data.frame with variant positions and 
the reagions to visualize.}
}
\description{
Generate igv batch screenshots script
}
