% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_master_table.R
\name{initiate_master_table}
\alias{initiate_master_table}
\title{Initiate master table}
\usage{
initiate_master_table(..., method_names)
}
\arguments{
\item{...}{VCF file addresses.}

\item{method_names}{Vector of strings of same length of number of input VCF files.}
}
\value{
A data.frame.
}
\description{
This function create the first columns of the master table. Those columns indicate
whether a method could call of not a variant. A master table is used to get
information form VCF files generated from different methods and compare to
ground-truth VCF files.
}
