% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splice_junction_analysis_table}
\alias{splice_junction_analysis_table}
\title{Generate data for splice-junction analysis using multiple master tables}
\usage{
splice_junction_analysis_table(
  ...,
  experiment_names,
  truth_names,
  method_names,
  output_method_names,
  variant_type,
  min_isoseq_coverage
)
}
\arguments{
\item{...}{1-length strings. Paths to master tables saved as RDS files.}

\item{experiment_names}{A vector of strings with length equal to the number of master
tables input in `...`. Names of the datasets for each data table.}

\item{truth_names}{A vector of strings. Names of the ground truth in each data table
especified in `...`.}

\item{method_names}{A vector of strings with length equal to the number of master
tables input in `...`. Names of the methods to compare.}

\item{output_method_names}{A vector of strings with length equal to the number of master
tables input in `...`. Names of the methods to be in the output.}

\item{variant_type}{A 1-length string. Possible values are "snp" or "indel".}

\item{min_isoseq_coverage}{A 1-length integer. The threshod value for the minimum
Iso-Seq read coverage.}
}
\value{
A list of two elements (`acc_sj` and `n_test`) to be used to draw a chart to
 analyse the variant-calling performance of sites near to and far from splice junctions.
 `acc_sj` stores the calculated performance measures, and `n_test` stores the number of
 observed variants in each case.
}
\description{
Generate data to plot variant performance of sites near to and far from splice junctions
  comparisons using multiple master table as facets.
}
